Move all the enums to their own module
We have a lot of enums in Airlock. In addition to reducing the size
of business_logic_layer.py, moving them to their own modules avoids
circular imports when various different modules need to use them.

Note: the one enum in old_api/schema.py is left where it is, as we
expect to replace old_api at some point in the (nearish?) future.
