+++ airlock/views/request.py
@@ -15,14 +15,8 @@ from django.views.decorators.vary import vary_on_headers
 from opentelemetry import trace
 
 from airlock import exceptions, permissions
-from airlock.business_logic import (
-    ROOT_PATH,
-    RequestFileType,
-    RequestFileVote,
-    RequestStatus,
-    Visibility,
-    bll,
-)
+from airlock.business_logic import ROOT_PATH, bll
+from airlock.enums import RequestFileType, RequestFileVote, RequestStatus, Visibility
 from airlock.file_browser_api import get_request_tree
 from airlock.forms import GroupCommentDeleteForm, GroupCommentForm, GroupEditForm
 from airlock.types import UrlPath
