+++ airlock/views/workspace.py
@@ -10,13 +10,15 @@ from django.views.decorators.vary import vary_on_headers
 from opentelemetry import trace
 
 from airlock import exceptions, permissions
-from airlock.business_logic import (
-    RequestFileType,
-    bll,
-)
+from airlock.business_logic import bll
+from airlock.enums import RequestFileType, WorkspaceFileStatus
 from airlock.file_browser_api import get_workspace_tree
-from airlock.forms import AddFileForm, FileTypeFormSet, MultiselectForm
-from airlock.types import UrlPath, WorkspaceFileStatus
+from airlock.forms import (
+    AddFileForm,
+    FileTypeFormSet,
+    MultiselectForm,
+)
+from airlock.types import UrlPath
 from airlock.views.helpers import (
     display_form_errors,
     display_multiple_messages,
