+++ tests/functional/test_code_pages.py
@@ -3,7 +3,7 @@ from urllib.parse import urlsplit
 import pytest
 from playwright.sync_api import expect
 
-from airlock.business_logic import RequestStatus
+from airlock.enums import RequestStatus
 from tests import factories
