+++ tests/local_db/test_data_access.py
@@ -1,8 +1,8 @@
 import pytest
 
 from airlock import exceptions
-from airlock.business_logic import (
-    AuditEvent,
+from airlock.business_logic import AuditEvent
+from airlock.enums import (
     AuditEventType,
     RequestStatus,
     Visibility,
